pyproject.toml
[project]
name = "code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adk>=0.0.5",
    "google-adk>=1.22.1",
    "litellm>=1.81.0",
    "loguru>=0.7.3",
]
